Improve REST semantic and refactor into modules
