+++ web-app/src/user/user.module.ts
@@ -1,0 +1,5 @@
+import { Module } from '@nestjs/common';
+
+@Module({
+})
+export class UserModule { }
